# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Assinatura Streaming",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/planos": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Listar todos os planos ativos",
                "operationId": "03e1c7c3c995df055f01241255080a06",
                "responses": {
                    "200": {
                        "description": "Lista de planos ativos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "string",
                                                "example": "prod_1234"
                                            },
                                            "nome": {
                                                "type": "string",
                                                "example": "Plano Mensal"
                                            },
                                            "prices": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "price_id": {
                                                            "type": "string",
                                                            "example": "price_abc123"
                                                        },
                                                        "amount": {
                                                            "type": "number",
                                                            "example": 99.9
                                                        },
                                                        "currency": {
                                                            "type": "string",
                                                            "example": "BRL"
                                                        },
                                                        "recorrencia": {
                                                            "type": "string",
                                                            "example": "Mensal"
                                                        },
                                                        "interval": {
                                                            "type": "string",
                                                            "example": "month"
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Planos"
                ],
                "summary": "Criar um novo plano",
                "operationId": "6806e9fec7343a10c76923aa8ecb5fc7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "amount"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Plano Premium"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 199.9
                                    },
                                    "interval": {
                                        "type": "string",
                                        "example": "month"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Plano criado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plano criado com sucesso!"
                                        },
                                        "id": {
                                            "type": "string",
                                            "example": "prod_abc123"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/planos/{id}/edit": {
            "get": {
                "tags": [
                    "Planos"
                ],
                "summary": "Exibe os detalhes de um plano para edição",
                "operationId": "4a409840ef1a8855b002dfc31aa04915",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do plano (product_id do Stripe)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "prod_abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do plano para edição",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "string",
                                            "example": "prod_abc123"
                                        },
                                        "nome": {
                                            "type": "string",
                                            "example": "Plano Premium"
                                        },
                                        "prices": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "price_id": {
                                                        "type": "string",
                                                        "example": "price_abc123"
                                                    },
                                                    "amount": {
                                                        "type": "number",
                                                        "example": 199.9
                                                    },
                                                    "currency": {
                                                        "type": "string",
                                                        "example": "BRL"
                                                    },
                                                    "interval": {
                                                        "type": "string",
                                                        "example": "month"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plano não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/planos/{id}": {
            "put": {
                "tags": [
                    "Planos"
                ],
                "summary": "Atualiza os detalhes de um plano",
                "operationId": "4ed4e6040671b71c56a4eb7e2f216701",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do plano (product_id do Stripe)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "prod_abc123"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "price_id",
                                    "amount",
                                    "interval"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "example": "Plano Anual"
                                    },
                                    "price_id": {
                                        "type": "string",
                                        "example": "price_abc123"
                                    },
                                    "amount": {
                                        "type": "number",
                                        "example": 299.9
                                    },
                                    "interval": {
                                        "type": "string",
                                        "example": "year"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Plano atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plano atualizado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Plano não encontrado"
                    },
                    "500": {
                        "description": "Erro ao atualizar plano"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Planos"
                ],
                "summary": "Desativar um plano",
                "operationId": "e9c256933050b9070173058869a3c711",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do plano (product_id do Stripe)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "prod_abc123"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Plano desativado com sucesso!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Plano deletado com sucesso!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}